refactor: more zsh changes

+++ main.sh
@@ -26,10 +26,11 @@ utils/setup.sh
 # ============================
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)" "" --unattended
 
-rm ~/.zshrc
-ln -sf "$SCRIPT_DIR"/common_config/.zshrc "$DEST"/.zshrc
 chsh -s $(which zsh)
 exec zsh
+
+[[ ! -e ~/.zshrc ]] && rm ~/.zshrc
+ln -sf common_config/zshrc ~/.zshrc
 
 # Restore current directory of user
 popd > /dev/null
